// src/app/models/Product.ts
export class Product {
  name: string;
  code: string;
  description: string;
  company: string;
  image?: string;
  unit: string;
  unitPrice?: number;
  category: string;
  hotsaleType: string;
  sellType: string;
  disponibility: boolean;
  hotsalePrice?: number;
  discountUnit?: number;
  discountPrice?: number;
  discountUnit2?: number;
  discountPrice2?: number;
  discountUnit3?: number;
  discountPrice3?: number;
}